+++ src/GetAudioEngine.php
@@ -6,7 +6,6 @@ namespace AKlump\ChangeAudio;
 use AKlump\ChangeAudio\Cache\CacheManager;
 use AKlump\ChangeAudio\Engine\EngineInterface;
 use AKlump\ChangeAudio\Engine\MacOSAudioDevicesEngine;
-use AKlump\ChangeAudio\Engine\SwitchAudioCommandEngine;
 use AKlump\ChangeAudio\Engine\SwitchAudioOSXEngine;
 
 class GetAudioEngine {
@@ -21,11 +20,10 @@ class GetAudioEngine {
    * @return EngineInterface|null The applicable audio engine or NULL if no engine applies.
    */
   public function __invoke(): ?EngineInterface {
-    /** @var MacOSAudioDevicesEngine|SwitchAudioCommandEngine|SwitchAudioOSXEngine[] $engines_by_priority The first that applies() === TRUE will be used. */
+    /** @var MacOSAudioDevicesEngine|SwitchAudioOSXEngine[] $engines_by_priority The first that applies() === TRUE will be used. */
     $engines_by_priority = [
       new MacOSAudioDevicesEngine(new CacheManager()),
       new SwitchAudioOSXEngine(),
-      new SwitchAudioCommandEngine(),
     ];
     foreach ($engines_by_priority as $engine) {
       if ($engine->applies()) {